Fix "roll" rotation.
Fix the rotation done for "roll" in PLRenderableElementBase (such as
the camera).  Previously the roll was done around the same axis as yaw,
which does not make sense.  Now pitch/yaw/roll will each rotate around
one of the main coordinate axes, so an arbitrary rotation can be
described by the three angles.