Support direct rotation matrix in PLCamera.
Add a method to set the camera rotation directly as a rotation matrix,
instead of via the Euler angles.  These two are mutually exclusive,
and applying one unsets / overwrites the other.

The method with Euler angles is still 100% the same as it was before,
so no existing code using it will break.  But it provides the option to
use a rotation matrix directly as well, which can help to circumvent
certain issues such as gimbal lock.